% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b1_samp.R
\name{b1_samp}
\alias{b1_samp}
\title{Sample beta1 given beta0 and sigma^2}
\usage{
b1_samp(draws, x, y, n)
}
\arguments{
\item{draws}{matrix containing at least one set of beta0 and sigma^2 as initial values}

\item{x}{covariate}

\item{y}{response variable}

\item{n}{number of measurements}
}
\value{
an updated beta1 value
}
\description{
Sample beta1 given beta0 and sigma^2
}
\details{
Sample one beta1 value from its full conditional distribution given beta0 and sigma^2
}
\examples{
data(simulated_data)
draws <- matrix(NA, 100, 3, dimnames = list(NULL, c("sigma2", "beta0", "beta1")))
draws[1, ] <- 1
b1_samp(draws=draws,x=simulated_data$x,y=simulated_data$y,n=1000)
}
\references{
Lesaffre, E. &Lawson, A. (2012). Bayesian Biostatistics. John Wiley & Sons, Ltd. Retrieved from https://onlinelibrary.wiley.com/doi/book/10.1002/9781119942412

Wakefield, J. (2013). Bayesian and Frequentist Regression Methods. Springer Series in Statistics. Retrieved from https://link.springer.com/book/10.1007/978-1-4419-0925-1
}
\author{
Meilin Jiang
}
