% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\docType{data}
\name{dat}
\alias{dat}
\title{Baby name data set}
\format{
A data frame with 67 rows and 5 variables:
\describe{
\item{year}{the year a baby was born}
\item{sex}{the gender of a baby}
\item{name}{the name of a baby}
\item{n}{the number of babies' names equal to 'name' column}
\item{prop}{the proportion of babies' name equal to 'name' column}
}
}
\source{
\url{https://github.com/hadley/babynames}
}
\usage{
dat
}
\description{
This is a data set containing 67 rows of female babies named as 'Sophia', who were born after year 1950.
}
\keyword{datasets}
