% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deterministic.R
\name{deterministic}
\alias{deterministic}
\title{Deterministic- or systematic scan Gibbs sampler}
\usage{
deterministic(draws, x, y, n, B)
}
\arguments{
\item{draws}{matrix containing one set of sigma2, beta0, and beta1 as initial values}

\item{x}{covariate}

\item{y}{response variable}

\item{n}{number of measurements}

\item{B}{number iterations (sampled values) desired}
}
\value{
a matrix of B vectors of sampled sigma2, beta0, and beta1 values
}
\description{
Deterministic- or systematic scan Gibbs sampler
}
\details{
The d=3 dimensions (sigma2, beta0, and beta1) are visited in a fixed order.
Sample B iterations of sigma2, beta0, and beta1 values in this order from their full conditional distributions
}
\examples{
data(simulated_data)
draws <- matrix(NA, 100, 3, dimnames = list(NULL, c("sigma2", "beta0", "beta1")))
draws[1, ] <- 1
deterministic(draws = draws,x=simulated_data$x,y=simulated_data$y,n=1000, B = 100)
}
\references{
Lesaffre, E. &Lawson, A. (2012). Bayesian Biostatistics. John Wiley & Sons, Ltd. Retrieved from https://onlinelibrary.wiley.com/doi/book/10.1002/9781119942412

Wakefield, J. (2013). Bayesian and Frequentist Regression Methods. Springer Series in Statistics. Retrieved from https://link.springer.com/book/10.1007/978-1-4419-0925-1
}
\author{
Meilin Jiang
}
