% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condPDF.R
\name{drho}
\alias{drho}
\title{Conditional log PDF of rho}
\usage{
drho(z, rho, log = FALSE, sum = FALSE)
}
\arguments{
\item{z}{vector length \code{n} of group membership}

\item{rho}{vector length \code{k} of prior group membership probability}

\item{log}{logical value whether to output log density or not. it defaults to \code{FALSE}}

\item{sum}{logical value whether to output summation of densities of observations. It defaults to \code{FALSE}}
}
\value{
vector length \code{n} of (log) density of given \code{rho} values when \code{sum=FALSE}. It produces integer that sums up all (log) densities of \code{rho}
}
\description{
Conditional log PDF of rho
}
