% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condPDF.R
\name{dSig}
\alias{dSig}
\title{Conditional log PDF of Sigma}
\usage{
dSig(theta, z, mu, Sigma, Omega, v_k, log = FALSE, sum = FALSE)
}
\arguments{
\item{theta}{matrix \verb{n x p} of observation mean}

\item{z}{vector length \code{n} of group membership}

\item{mu}{matrix \verb{k x p} of group mean}

\item{Sigma}{matrix \verb{p x p x k} of within-group variance}

\item{Omega}{matrix \verb{p x p x k} of between row precision}

\item{v_k}{vector length \code{k} of prior contribution to each cluster}

\item{log}{logical value whether to output log density or not. it defaults to \code{FALSE}}

\item{sum}{logical value whether to output summation of densities of observations. It defaults to \code{FALSE}}
}
\value{
vector length \code{n} of (log) density of given \code{Sigma} values when \code{sum=FALSE}. It produces integer that sums up all (log) densities of \code{Sigma}
}
\description{
Conditional log PDF of Sigma
}
