% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condPDF.R
\name{dtheta}
\alias{dtheta}
\title{Conditional log PDF of theta}
\usage{
dtheta(y, theta, z, mu, Sig, V, log = FALSE, sum = FALSE)
}
\arguments{
\item{y}{matrix \verb{n x p} of observation}

\item{theta}{matrix \verb{n x p} of observation mean}

\item{z}{vector length \code{n} of group membership}

\item{mu}{matrix \verb{k x p} of group mean}

\item{Sig}{matrix \verb{p x p x k} of within-group variance}

\item{V}{matrix \verb{p x p x n} of observation variance}

\item{log}{logical value whether to output log density or not. it defaults to \code{FALSE}}

\item{sum}{logical value whether to output summation of densities of observations. It defaults to \code{FALSE}}
}
\value{
vector length \code{n} of (log) density of given \code{theta} values when \code{sum=FALSE}. It produces integer that sums up all (log) densities of \code{theta}
}
\description{
Conditional log PDF of theta
}
