% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condPDF.R
\name{dz}
\alias{dz}
\title{Conditional log PDF of z}
\usage{
dz(z, rho, theta, mu, Sigma, log = FALSE, sum = FALSE)
}
\arguments{
\item{z}{vector length \code{n} of group membership}

\item{rho}{vector length \code{n} of prior group membership probability}

\item{theta}{matrix \verb{n x p} of observation mean}

\item{mu}{matrix \verb{k x p} of group mean}

\item{Sigma}{matrix \verb{p x p x k} of within-group variance}

\item{log}{logical value whether to output log density or not. it defaults to \code{FALSE}}

\item{sum}{logical value whether to output summation of densities of observations. It defaults to \code{FALSE}}
}
\value{
vector length \code{n} of (log) density of given \code{z} values when \code{sum=FALSE}. It produces integer that sums up all (log) densities of \code{z}
}
\description{
Conditional log PDF of z
}
