% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gibbs-package.R
\docType{package}
\name{Gibbs-package}
\alias{Gibbs}
\alias{Gibbs-package}
\title{Gibbs: Gibbs Sampler for model-based clustering}
\description{
Use Gibbs Sampler to hierarchical Gaussian mixture model for clustering
}
\author{
\strong{Maintainer}: Jeonghun Jeon \email{j2jeon@uwaterloo.ca} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
