% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gibbs.R
\name{gibbs}
\alias{gibbs}
\title{Gibbs sampling for simple linear regression model}
\usage{
gibbs(X, Y, n)
}
\arguments{
\item{X}{Input values used as independent variable in the linear model}

\item{Y}{Observed values used as dependent variable in the linear model}

\item{n}{Number of input values/number of rows in the data set used to fit the linear model}
}
\value{
GibbsSampleResult: Sampling results for Intercept, Slope and sigma2 (sampling for 30000 times)

GibbsSampleSummary: Summary of the sampling results (Mean, SD, 2.5\% quantile, 97.5\% quantile)

SD_Comparison: Comparison between the standard deviation of the posterior samples and the standard deviation of OLS

Plot of X and Y with fitted line from OLS

Histogram of the posterior slope

Plot of X and Y with the fitted line of Intercept mean and Slope mean from sampling
}
\description{
Gibbs sampling for simple linear regression model with one parameter and one outcome
}
\details{
Gibbs sampling or a Gibbs sampler is a Markov chain Monte Carlo (MCMC) algorithm for obtaining a
sequence of observations which are approximated from a specified multivariate probability
distribution. Gibbs sampling is commonly used as a means of statistical inference, especially
Bayesian inference. From which we can get the approximate joint posterior distribution and marginal
posterior distribution.
}
\author{
Luer Zhong
}
