% Generated by roxygen2 (4.0.0): do not edit by hand
\name{gibbsampler}
\alias{gibbsampler}
\title{A Gibb Sampler}
\usage{
gibbsampler(seed = sample(1:4, 1))
}
\arguments{
\item{seed}{Enter a random number or one is selected for you.}
}
\description{
This function samples gibbs.
}
\examples{
gibbsampler()
}
\keyword{gibbs}

