% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GibbsSampler.R
\name{sampling_mu}
\alias{sampling_mu}
\title{Sampling method for within-group sample mean.}
\usage{
sampling_mu(z, theta, mu, Sigma)
}
\arguments{
\item{z}{vector length \code{n} of observation's cluster allocation}

\item{theta}{matrix \verb{n x p} of observation means}

\item{mu}{matrix \verb{k x p} of mean values for each cluster}

\item{Sigma}{matrix \verb{p x p x k} of variances for each cluster}
}
\value{
matrix \verb{k x p} of sampled mu
}
\description{
Sampling method for within-group sample mean.
}
\details{
\code{mu_k} has conditional distribution \verb{N(\\hat\\theta_k,\\Sigma_k/N_k)} where \verb{N_k=\\sum_\{i=1\}^\{N\}I(z_i=k)}. Sampling is done using \code{rmNorm()} from \code{mniw} package.
}
\references{
Martin Lysy and Bryan Yates (2019). mniw: The Matrix-Normal Inverse-Wishart Distribution. R package version 1.0. https://CRAN.R-project.org/package=mniw
}
