% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GibbsSampler.R
\name{sampling_rho}
\alias{sampling_rho}
\title{Sampling method for probability vector for group membership.}
\usage{
sampling_rho(z, k)
}
\arguments{
\item{z}{vector length \code{n} of observation's cluster allocation}

\item{k}{integer of number of clusters in the data}
}
\value{
vector length \code{k} of prior distribution for clusters
}
\description{
Sampling method for probability vector for group membership.
}
\details{
\code{rho_k} has conditional distribution \verb{Dirichlet(\\alpha)}, where \verb{\\alpha=(N_k+1)}.
}
