% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GibbsSampler.R
\name{sampling_sigma}
\alias{sampling_sigma}
\title{Sampling method for within-group variance-covariance matrix}
\usage{
sampling_sigma(z, mu, Omega, theta, v_k)
}
\arguments{
\item{z}{vector length \code{n} of observation's cluster allocation}

\item{mu}{matrix \verb{n x p} of mean values for each cluster}

\item{Omega}{matrix \verb{p x p x k} of between row precisions}

\item{theta}{matrix \verb{n x p} of observation means}

\item{v_k}{vector length \code{k} of prior contribution to each cluster}
}
\value{
matrix \verb{p x p x k} of sampled sigma
}
\description{
Sampling method for within-group variance-covariance matrix
}
\details{
\code{Sigma_k} has conditional distribution \verb{InvWish(\\Omega_k + \\sum_\{i:I(z_i=k)\}(\\theta_i-\\mu_k)(\\theta_i-\\mu_k)',N_k+v_k)}. Sampling is done using \code{riwish()} function from \code{mniw} package.
}
\references{
Martin Lysy and Bryan Yates (2019). mniw: The Matrix-Normal Inverse-Wishart Distribution. R package version 1.0. https://CRAN.R-project.org/package=mniw
}
