% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GibbsSampler.R
\name{sampling_z}
\alias{sampling_z}
\title{Sampling method for probability matrix for group membership and assign group membership based on these probabilities}
\usage{
sampling_z(theta, rho, mu, Sigma, prior_z, prior_lambda)
}
\arguments{
\item{theta}{matrix \verb{n x p} of parameter values for observation x}

\item{rho}{vector length \code{k} of prior distribution for clusters}

\item{mu}{matrix \verb{k x p} of mean values for each clusters}

\item{Sigma}{matrix \verb{p x p x k} of variance for each clusters}

\item{prior_z}{vector length \code{n} of clusters allocation, which will be returned instead of newly sampled cluster allocation when at least one N_k = 0}

\item{prior_lambda}{matrix \verb{n x k} of posterior probabilities, which will be returned instead of newly sampled posteriro probabilities when at least one N_k = 0}
}
\value{
matrix \verb{n x k} of observations' posterior probabilities

vector length \code{n} of observations' pooled cluster allocation
}
\description{
Sampling method for probability matrix for group membership and assign group membership based on these probabilities
}
\details{
Group membership \code{z} is chosen based on maximum a posteriori criteria. If there is at least 1 group that has not been assigned, it would not update \code{lambda} and \code{z}.
}
