% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gibbs.R
\name{Gibbs}
\alias{Gibbs}
\title{Gibbs Sampling for Simple Linear Regression}
\usage{
Gibbs(x, y, method = "deterministic", sigma2_initial = NA,
  beta0_initial = NA, beta1_initial = NA, B = 1000, seed = 123)
}
\arguments{
\item{x}{covariate}

\item{y}{response variable}

\item{method}{a choice of the following Gibbs sampling methods: deterministic, random, reversible, or block}

\item{sigma2_initial}{initial value for sigma^2. MLE is used if not supplied}

\item{beta0_initial}{initial value for beta0. MLE is used if not supplied}

\item{beta1_initial}{initial value for beta1. MLE is used if not supplied}

\item{B}{number iterations (sampled values) desired}

\item{seed}{seed set for random number generation}
}
\value{
a matrix of B vectors of sampled sigma2, beta0, and beta1 values
}
\description{
Gibbs Sampling for Simple Linear Regression
}
\details{
Use Gibbs sampler to sample parameters of linear regression (sigma2, beta0, and beta1) from full conditionals to determine their posterior distributions.
Sample B iterations of sigma2, beta0, and beta1 values from their full conditional distributions.
}
\examples{
data(simulated_data)
draws <- Gibbs(x=simulated_data$x,y=simulated_data$y,
sigma2_initial=1, beta0_initial=1,beta1_initial=1,method="block",B=1000)
apply(draws, 2, mean)
}
\references{
Lesaffre, E. &Lawson, A. (2012). Bayesian Biostatistics. John Wiley & Sons, Ltd. Retrieved from https://onlinelibrary.wiley.com/doi/book/10.1002/9781119942412

Wakefield, J. (2013). Bayesian and Frequentist Regression Methods. Springer Series in Statistics. Retrieved from https://link.springer.com/book/10.1007/978-1-4419-0925-1
}
\author{
Meilin Jiang
}
